Makefile enhancements.
The original Makefile built everything unconditionally; it also built
everything in the same directory as the source. Additionally, I had to
point it at the correct JDK path for my system and add a -fPIC argument
to gcc to avoid a linker warning.

All of these issues have been fixed in this rewrite, though it is obviously
not an optimal Makefile. (I haven't used Make in a long time and just
wanted to get something working quickly without having to spend too much
time reacquainting myself with the structure of Makefile rules. I'll fix
this later.