Modified test program to search for my stylus device.
In the past, one typically added the tablet input devices in the X11
config file, and traditional names were 'stylus' and 'eraser'. Ubuntu
(at least) has done away with the xorg.conf file, and the default
names for input devices are determined by their drivers, and do not
follow the conventions described in the XDeviceInfo man page.

At some point, this hard-coded value should be moved out to a config
file or dynamically generated, but I needed a quick fix for the moment.